+++ src/js/d2c-enhancements.js
@@ -1182,7 +1182,7 @@
   }
 
   // src/js/modules/build-date.js
-  var BUILD_DATE = "2026-02-24 13:45";
+  var BUILD_DATE = "2026-02-24 13:51";
 
   // src/js/modules/help-button.js
   function buildHelpButton() {